backend/requirements.txt
# SmartScheduler MVP 의존성
# 작성일: 2025-11-26

# Core - Gemini AI
google-generativeai>=0.8.0

# Discord Bot (Phase 5)
discord.py>=2.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Note: python-dotenv는 사용하지 않음 (config.py에서 직접 파싱)
